+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0
+
+### Added
+- Added bindIf(), bindSingletonIf(), bindSharedIf() to Container::class.
+- Added getAbstract() to Definition::class.
+
 ## 1.0.0
 
 ### Added